Fix not extract the not all words and means
Before, Some words and means not look good extraction.
example, "::((" <- that weird result.
That problem made from the revers methond that exist for remove unnecessary tags.
But I've found it is not right way.
So, Remove that revers method and append all of it and insert '@'
charactor between original string.
Signed-off-by: Seokho Song <0xdevssh@gmail.com>